surfaceflinger: Let external display roratable
By default, a landscape external display never rorate itself to portrait
while displaying portrait contents (or while ro.sf.hwrotation is set to 90
or 270). Instead, it shrinks places the portrait content inside with black
background around.

This commit make it possible to rotate the external display 90deg or 270deg,
just like it is an built-in primary display.

To enable this feature, set:

    TARGET_ROTATES_EXTERNAL_DISPLAY:= true

Change-Id: I0b30da6e86dd46defbb4a126276f3755961f99de